+++ app/models/note.rb
@@ -8,7 +8,7 @@
 #  notable_tag  :string
 #  created_at   :datetime         not null
 #  updated_at   :datetime         not null
-#  style        :string           default("original"), not null
+#  style        :string           default("blue"), not null
 #  body         :text
 #
 
@@ -23,7 +23,7 @@ class Note < ApplicationRecord
 
   cattr_accessor :default_style,
                  instance_accessor: false,
-                 default: 'original'
+                 default: 'blue'
   cattr_accessor :style_labels,
                  instance_accessor: false,
                  default: {