+++ db/migrate/20260713141538_change_default_note_style.rb
@@ -1,0 +1,5 @@
+class ChangeDefaultNoteStyle < ActiveRecord::Migration[8.0]
+  def change
+    change_column_default :notes, :style, from: 'original', to: 'blue'
+  end
+end